# swagger-api-doc.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/user/admin/updateUser": {
      "put": {
        "tags": [
          "user-controller"
        ],
        "operationId": "update",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/api/task/user/updateComplete": {
      "put": {
        "tags": [
          "task-controller"
        ],
        "operationId": "setTaskComplete",
        "parameters": [
          {
            "name": "taskId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/task/admin/updateTask": {
      "put": {
        "tags": [
          "task-controller"
        ],
        "operationId": "update_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          }
        }
      }
    },
    "/api/task/admin/removeUserFromTask/{taskId}": {
      "put": {
        "tags": [
          "task-controller"
        ],
        "operationId": "removeUserFromTask",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTableResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/task/admin/assignUser{taskId}/{userId}": {
      "put": {
        "tags": [
          "task-controller"
        ],
        "operationId": "assignUserToTask",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTableResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/comment/admin/updateComment": {
      "put": {
        "tags": [
          "comment-controller"
        ],
        "operationId": "update_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/admin/createUser": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/api/task/admin/createTask": {
      "post": {
        "tags": [
          "task-controller"
        ],
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          }
        }
      }
    },
    "/api/comment/user/commentMyTask": {
      "post": {
        "tags": [
          "comment-controller"
        ],
        "operationId": "userCommentOnTask",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddUserComment"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommentsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/comment/admin/createComment": {
      "post": {
        "tags": [
          "comment-controller"
        ],
        "operationId": "create_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/admin/allUserList": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "operationId": "getAllUserList",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/task/user/allTaskList/{assignee}": {
      "get": {
        "tags": [
          "task-controller"
        ],
        "operationId": "getAllUserTaskList",
        "parameters": [
          {
            "name": "assignee",
            "in": "path",
            "required": true,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskTableResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/task/admin/allTaskList": {
      "get": {
        "tags": [
          "task-controller"
        ],
        "operationId": "getAllTaskList",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/comment/user/userCommentList/{userId}": {
      "get": {
        "tags": [
          "comment-controller"
        ],
        "operationId": "getAllUserCommentList",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentsResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/comment/admin/allCommentList": {
      "get": {
        "tags": [
          "comment-controller"
        ],
        "operationId": "getAllCommentList",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Comment"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/user/admin/deleteUser/{id}": {
      "delete": {
        "tags": [
          "user-controller"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "minimum": 2,
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/task/admin/deleteTask/{id}": {
      "delete": {
        "tags": [
          "task-controller"
        ],
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UpdateUserRequest": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "minimum": 2,
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "isAdmin": {
            "type": "boolean"
          },
          "active": {
            "type": "boolean"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN"
            ]
          },
          "admin": {
            "type": "boolean"
          }
        }
      },
      "UpdateTaskRequest": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string"
          },
          "description": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "assignee": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "TaskTableResponse": {
        "type": "object",
        "properties": {
          "task_id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "task_status": {
            "type": "string"
          },
          "task_assignee": {
            "type": "integer",
            "format": "int64"
          },
          "err": {
            "type": "string"
          }
        }
      },
      "UpdateCommentRequest": {
        "required": [
          "comment",
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "comment": {
            "maxLength": 120,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "comment": {
            "type": "string"
          },
          "userId": {
            "$ref": "#/components/schemas/User"
          },
          "taskId": {
            "$ref": "#/components/schemas/Task"
          }
        }
      },
      "CreateUserRequest": {
        "required": [
          "active",
          "email",
          "isAdmin",
          "name",
          "password"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "isAdmin": {
            "type": "boolean"
          },
          "active": {
            "type": "boolean"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "CreateTaskRequest": {
        "required": [
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string"
          },
          "description": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "AddUserComment": {
        "required": [
          "TaskId",
          "comment",
          "userId"
        ],
        "type": "object",
        "properties": {
          "comment": {
            "maxLength": 120,
            "minLength": 0,
            "type": "string"
          },
          "TaskId": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "CommentsResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "comment": {
            "type": "string"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "taskId": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "err": {
            "type": "string"
          }
        }
      },
      "CreateCommentRequest": {
        "required": [
          "TaskId",
          "comment"
        ],
        "type": "object",
        "properties": {
          "comment": {
            "maxLength": 120,
            "minLength": 0,
            "type": "string"
          },
          "TaskId": {
            "type": "integer",
            "format": "int64"
          }
        }
      }
    }
  }
}